# openapi.yaml
openapi: 3.0.0
info:
  title: TradeOn API
  version: 1.0.0
  description: 'TradeOn은 사용자 간 물물교환 기반의 중고 거래 플랫폼입니다.


    이 API 문서는 다음과 같은 기능들을 포함합니다:

    - 사용자 회원가입, 로그인 및 정보 조회

    - 교환 요청 및 거래 이력 관리

    - 사용자 리뷰 시스템

    - 가치 평가 게시판

    - 하이브리드 검색기능

    - 게시물 등록/삭제/조회

    - 채팅 및 거래 약속

    - 거래 데이터 분석 및 시각화


    JWT 인증이 필요한 API의 경우 Authorize 버튼을 통해 토큰을 입력해주세요.'
servers:
- url: https://673e-121-140-48-141.ngrok-free.app
  description: 로컬
...
paths:
...
  /api/v1/search:
    get:
      tags:
        - Search
      summary: 게시글 키워드 검색
      description: |
        title 및 category 파라미터를 AND 조건으로 검색합니다.  
        두 파라미터 중 하나는 필수이며, 둘 다 없으면 400 에러를 반환합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
          description: 검색할 게시글 제목 키워드
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: 검색할 카테고리
      responses:
        '400':
          description: title 또는 category 파라미터가 필요합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: 검색 결과 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  count:
                    type: integer
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchPostItem'
  /api/v1/search/advanced:
    get:
      tags:
        - Search
      summary: 제목 기반 고급 게시글 검색
      description: 입력한 제목(title)을 기반으로 하이브리드 검색을 수행하여 유사 게시글을 반환합니다.
      parameters:
        - name: title
          in: query
          required: true
          description: 검색할 게시글 제목
          schema:
            type: string
            example: 에어팟 프로 중고 팝니다
      responses:
        "200":
          description: 검색 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvancedSearchResponse'
        "400":
          description: 잘못된 요청 (title 없음)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    AdvancedSearchResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        count:
          type: integer
          example: 2
        posts:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                description: 게시글 ID
                example: 123
              title:
                type: string
                description: 게시글 제목
                example: 에어팟 프로 중고 팝니다
              user_id:
                type: integer
                description: 게시글 작성자 ID
                example: 1
              category:
                type: string
                description: 게시글 카테고리
                example: 전자기기
              thumbnail_image_url:
                type: string
                description: 게시글 썸네일 이미지 URL
                example: https://example.com/thumb.jpg
              created_at:
                type: string
                format: date
                description: 게시글 생성일 (YYYY-MM-DD)
                example: 2024-06-01
            required:
              - id
              - title
              - user_id
              - category
              - thumbnail_image_url
              - created_at
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        msg:
          type: string
          example: 오류 메시지입니다.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []